Move DTO classes into respective packages
