+++ src/main/java/bank/connect/tech/controller/AccountController.java
@@ -1,7 +1,7 @@
 package bank.connect.tech.controller;
 
-import bank.connect.tech.dto.AccountCreateDTO;
-import bank.connect.tech.dto.AccountUpdateDTO;
+import bank.connect.tech.dto.create.AccountCreateDTO;
+import bank.connect.tech.dto.update.AccountUpdateDTO;
 import bank.connect.tech.model.Account;
 import bank.connect.tech.response.SuccessResponse;
 import bank.connect.tech.service.AccountService;
