+++ src/main/java/bank/connect/tech/dto/update/AccountUpdateDTO.java
@@ -1,4 +1,4 @@
-package bank.connect.tech.dto;
+package bank.connect.tech.dto.update;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
