+++ src/main/java/bank/connect/tech/dto/update/BillUpdateDTO.java
@@ -1,4 +1,4 @@
-package bank.connect.tech.dto;
+package bank.connect.tech.dto.update;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.time.LocalDate;
